Authority adjustment in router for admin rights / Comment report routing and view done

+++ index.php
@@ -24,22 +24,25 @@ if (isset($_GET['action'])) {
 	else if ($_GET['action'] == 'comment_confirm'){
 		commentConfirm();
 	}
-	else if ($_GET['action'] == 'admin_panel'){
+	else if ($_GET['action'] == 'admin_panel' && $_SESSION['authority'] == 2){
 		adminPanel();
 	}
-	else if ($_GET['action'] == 'add_entry'){
+	else if ($_GET['action'] == 'add_entry' && $_SESSION['authority'] == 2){
 		addEntry();
 	}
-	else if ($_GET['action'] == 'add_entry_confirm'){
+	else if ($_GET['action'] == 'add_entry_confirm' && $_SESSION['authority'] == 2){
 		if (isset($_GET['id'])){
 			updateEntryConfirm();
 		}
 		else{
 			addEntryConfirm();}
 	}
-	else if ($_GET['action'] == 'delete_entry'){
+	else if ($_GET['action'] == 'delete_entry' && $_SESSION['authority'] == 2){
 		deleteEntryConfirm();
 
+	}
+	else if ($_GET['action'] == 'report_comment'){
+		
 	}
 	else {
 		echo 'Lien invalide';
